rename GameVideo file
